refactor(crm): add entrypoints, deduplicate query, redirect imports
T1 — Create module entrypoints:
  - Added server.ts (exports crmService)
  - Added index.ts (comment-only, no client exports yet)
  - CRM now follows the same module boundary pattern as all other modules

T2 — Redirect consumers to entrypoint:
  - frank/frank.service.ts: crm/crm.service -> crm/server
  - atendimento/freight-quote.service.ts: crm/crm.service -> crm/server
  - atendimento/message.service.ts: crm/crm.service -> crm/server

T3 — Extract findActiveOpportunity helper:
  - Deduplicated the 'find active opportunity by tenant+customer' query
    that was identically repeated in projectMessageActivity and
    syncSimulationToQuote (10 lines each)
  - Single private helper, same query, same behavior

Zero functional changes. No schema changes. No business logic changes.

QG Results:
- typecheck: PASS
- lint:env: PASS
- lint:pii: PASS
- db:verify: PASS (no drift)
- routes:inventory: PASS (342 routes)
- routes:verify: PASS (342 routes)
- routes:verify-security: PASS (160 routes)
- test:ci: PASS (917 tests, 187 files)
- check:boundaries: PASS (1266 files)
- build: PASS
- release:check: RELEASE_CHECK_PASS
- npm audit: 0 vulnerabilities
