+++ src/modules/crm/index.ts
@@ -1,0 +1,1 @@
+// CRM module — client-safe exports (types only, no server code)
