+++ src/modules/crm/server.ts
@@ -1,0 +1,1 @@
+export { crmService } from './crm.service';
